==> R/2_all_generics.R <==
setGeneric(
  name = "database_info",
  def = function(object, ...)
    standardGeneric(f = "database_info")
)


setGeneric(
  name = "pathway_id",
  def = function(object, ...)
    standardGeneric(f = "pathway_id")
)

setGeneric(
  name = "pathway_name",
  def = function(object, ...)
    standardGeneric(f = "pathway_name")
)


setGeneric(
  name = "describtion",
  def = function(object, ...)
    standardGeneric(f = "describtion")
)


setGeneric(
  name = "pathway_class",
  def = function(object, ...)
    standardGeneric(f = "pathway_class")
)

setGeneric(
  name = "gene_list",
  def = function(object, ...)
    standardGeneric(f = "gene_list")
)

setGeneric(
  name = "compound_list",
  def = function(object, ...)
    standardGeneric(f = "compound_list")
)


setGeneric(
  name = "reference_list",
  def = function(object, ...)
    standardGeneric(f = "reference_list")
)


setGeneric(
  name = "related_disease",
  def = function(object, ...)
    standardGeneric(f = "related_disease")
)


setGeneric(
  name = "related_module",
  def = function(object, ...)
    standardGeneric(f = "related_module")
)
