==> R/5_data.R <==
#' #' @title KEGG Compound Database
#' #' @description A dataset containing compound information from KEGG.
#' #' @format A data frame with KEGG compound details.
#' #' @docType data
#' #' @usage data(kegg_compound_database)
#' #' @keywords datasets
#' "kegg_compound_database"
#' 
#' #' @title HMDB Compound Database
#' #' @description A dataset containing compound information from HMDB.
#' #' @format A data frame with HMDB compound details.
#' #' @docType data
#' #' @usage data(hmdb_compound_database)
#' #' @keywords datasets
#' "hmdb_compound_database"
#' 
#' #' @title KEGG Human Pathway Database
#' #' @description A dataset containing KEGG pathways for Homo sapiens.
#' #' @format A data frame with KEGG pathway data.
#' #' @docType data
#' #' @usage data(kegg_hsa_pathway)
#' #' @keywords datasets
#' "kegg_hsa_pathway"
#' 
#' 
#' #' @title HMDB Pathway Database
#' #' @description A dataset containing HMDB pathways for Homo sapiens.
#' #' @format A data frame with KEGG pathway data.
#' #' @docType data
#' #' @usage data(hmdb_pathway)
#' #' @keywords datasets
#' "hmdb_pathway"
#' 
#' #' @title Reactome Human Pathway Database
#' #' @description A dataset containing Reactome pathways for Homo sapiens.
#' #' @format A data frame with Reactome pathway data.
#' #' @docType data
#' #' @usage data(reactome_hsa_pathway)
#' #' @keywords datasets
#' "reactome_hsa_pathway"
#' 
#' #' @title WikiPathway Human Pathway Database
#' #' @description A dataset containing WikiPathways pathways for Homo sapiens.
#' #' @format A data frame with WikiPathways pathway data.
#' #' @docType data
#' #' @usage data(wikipathway_hsa_pathway)
#' #' @keywords datasets
#' "wikipathway_hsa_pathway"
#' 
#' 
#' #' @title HMDB Query ID List
#' #' @description A dataset containing HMDB compound IDs for querying.
#' #' @author Xiaotao Shen (\email{shenxt1990@outlook.com})
#' #' @format A character vector of HMDB compound IDs.
#' "query_id_hmdb"
#' 
#' #' @title KEGG Query ID List
#' #' @description A dataset containing KEGG compound IDs for querying.
#' #' @author Xiaotao Shen (\email{shenxt1990@outlook.com})
#' #' @format A character vector of KEGG compound IDs.
#' "query_id_kegg"
#' 
#'
