==> R/6_global_variables.R <==


globalVariables(
  names = c(
    "CAS.ID",
    "Compound.name",
    "Formula",
    "HMDB.ID",
    "KEGG.ID",
    "Lab.ID",
    "RT",
    "Submitter",
    "compound_class",
    "data",
    "hmdb_pathway",
    "keggMS1database",
    "kegg_hsa_pathway",
    "mz",
    "mz.neg",
    "mz.pos",
    "n",
    "name",
    ".",
    "hmdbMS1Database",
    "x_value",
    "x",
    "pathway_name",
    "mapped_id",
    "jaccard_index",
    "geom_node_text",
    "y",
    "p_value",
    "mapped_number",
    "all_number",
    "object",
    "hmdb_compound_database",
    "x_axis",
    "cutoff",
    "",
    "temp.ID",
    "query_type",
    "Adduct",
    "Compound_name",
    "HMDB_ID" ,
    "KEGG_ID",
    "Total_metabolite_id",
    "degree",
    "ecdf",
    "edge_class",
    "from",
    "from_compound_KEGG_ID",
    "isotope",
    "mzError.ppm",
    "node_class",
    "peakIndex",
    "polarity",
    "rgamma",
    "rt",
    "rtError.s",
    "score",
    "to",
    "to_compound_KEGG_ID",
    "variable_id",
    "metabolic_module_name",
    "module"
  )
)
